Fixed structural error: now device's address book opens correctly
