+++ demo/src/main/java/com/eis/geoCalendar/demo/ContactsActivity.java
@@ -2,13 +2,99 @@ package com.eis.geoCalendar.demo;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Intent;
+import android.database.Cursor;
+import android.net.Uri;
 import android.os.Bundle;
+import android.provider.ContactsContract;
+import android.view.View;
+import android.widget.Button;
+import android.widget.LinearLayout;
+import android.widget.TextView;
+import android.widget.Toast;
+
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+
+/**
+ * Class to pick contacts from the device's address book and display them as a contacts list.
+ *
+ * @author Alessandra Tonin
+ * @since 23/02/2020
+ */
 
 public class ContactsActivity extends AppCompatActivity {
+
+    //This code will be used to start an activity and then passed to onActivityResult();
+    //It's useful to differentiate between different results.
+    private static final int PICK_CONTACT = 1;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_contacts);
+
+        final FloatingActionButton button = findViewById(R.id.addContactButton);
+        button.setOnClickListener(new View.OnClickListener() {
+            public void onClick(View v) {
+                openAddressBook(v);
+            }
+        });
     }
+
+    /**
+     * Method to open the system address book
+     *
+     * @param view The view calling the method
+     */
+    public void openAddressBook(View view) {
+        Intent intent = new Intent(Intent.ACTION_PICK, ContactsContract.Contacts.CONTENT_URI);
+        startActivityForResult(intent, PICK_CONTACT);
+    }
+
+
+    /**
+     * Method to handle the picked contact
+     *
+     * @param requestCode The code of the request
+     * @param resultCode  The result of  the request
+     * @param data        The data of the result
+     */
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+        if (requestCode == PICK_CONTACT) {
+            if (resultCode == RESULT_OK) {
+                Uri contactData = data.getData();
+                String number = "";
+                Cursor cursor = getContentResolver().query(contactData, null, null, null, null);
+                cursor.moveToFirst();
+                String hasPhone = cursor.getString(cursor.getColumnIndexOrThrow(ContactsContract.Contacts.HAS_PHONE_NUMBER));
+                String contactId = cursor.getString(cursor.getColumnIndexOrThrow(ContactsContract.Contacts._ID));
+                if (hasPhone.equals("1")) {
+                    Cursor phones = getContentResolver().query
+                            (ContactsContract.CommonDataKinds.Phone.CONTENT_URI, null,
+                                    ContactsContract.CommonDataKinds.Phone.CONTACT_ID
+                                            + " = " + contactId, null, null);
+                    while (phones.moveToNext()) {
+                        number = phones.getString(phones.getColumnIndex
+                                (ContactsContract.CommonDataKinds.Phone.NUMBER)).replaceAll("[-() ]", "");
+                    }
+                    phones.close();
+
+                    //Put the number in the scrollView
+                    LinearLayout linearLayout = (LinearLayout) this.findViewById(R.id.linLayout);
+                    //Create a new TextView with contact number
+                    TextView contactName = new TextView(this);
+                    contactName.setText(number);
+                    //Adding textView to LinearLayout
+                    linearLayout.addView(contactName);
+
+                } else {
+                    Toast.makeText(getApplicationContext(), "This contact has no phone number", Toast.LENGTH_LONG).show();
+                }
+                cursor.close();
+            }
+        }
+    }
+
 }
